Move a few functions to nodes.py
